assignment_3/Auto-Backup System Using Google Drive API/requirements.txt
flask
watchdog
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
